+++ fabric/src/main/java/io/github/sakurawald/module/initializer/echo/send_title/SendTitleInitializer.java
@@ -6,9 +6,6 @@ import io.github.sakurawald.core.command.annotation.CommandNode;
 import io.github.sakurawald.core.command.annotation.CommandRequirement;
 import io.github.sakurawald.core.command.annotation.CommandSource;
 import io.github.sakurawald.module.initializer.ModuleInitializer;
-import net.minecraft.network.packet.s2c.play.SubtitleS2CPacket;
-import net.minecraft.network.packet.s2c.play.TitleFadeS2CPacket;
-import net.minecraft.network.packet.s2c.play.TitleS2CPacket;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.Text;
@@ -33,13 +30,11 @@ public class SendTitleInitializer extends ModuleInitializer {
         int $stayTicks = stayTicks.orElse(70);
         int $fadeOutTicks = fadeOutTicks.orElse(20);
 
-        Text mainTitleComponent = TextHelper.getTextByValue(player, $mainTitle);
-        Text subTitleComponent = TextHelper.getTextByValue(player, $subTitle);
+        Text mainTitleText = TextHelper.getTextByValue(player, $mainTitle);
+        Text subTitleText = TextHelper.getTextByValue(player, $subTitle);
 
-        player.networkHandler.sendPacket(new TitleFadeS2CPacket($fadeInTicks, $stayTicks, $fadeOutTicks));
-        player.networkHandler.sendPacket(new TitleS2CPacket(mainTitleComponent));
-        player.networkHandler.sendPacket(new SubtitleS2CPacket(subTitleComponent));
-
+        TextHelper.sendTitle(player, $fadeInTicks, $stayTicks, $fadeOutTicks, mainTitleText, subTitleText);
         return CommandHelper.Return.SUCCESS;
     }
+
 }