+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
 - [top_chunks] feature: new GUI interface to display top chunks. (Include a searching function)
+- [command_cooldown] feature: use `title` interface to display the cooldown message.
+